# swagger.yml
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
openapi: 3.0.0
info:
  title: Node.js API with Swagger 🚀
  description: Documentation for Node.js API with MongoDB and RabbitMQ 🐇
  version: 1.0.0
servers:
  - url: http://localhost:5000/
    description: Local development server 🔧
paths:
...
  /api/queue/enqueue:
    post:
      summary: Enqueue a task 📥
      description: Enqueues a task to the user's specific queue in RabbitMQ.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                task:
                  type: string
                  description: The task to enqueue
                  example: "Example task for user"
      responses:
        '200':
          description: Task enqueued successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Task enqueued
        '401':
          description: Unauthorized ❌
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: No token provided
        '500':
          description: Internal Server Error ❌
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Failed to enqueue task
      security:
        - bearerAuth: []
...